# RecordPoint.Connectors.SDK.Client/Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Records365 vNext Connector API",
    "description": "This is the Records365 vNext Connector API Home",
    "termsOfService": "See Legal"
  },
  "basePath": "/connector",
  "paths": {
...
    "/api/Notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Gets a collection of notifications that are awaiting processing and acknowledgement by the connector.",
        "operationId": "_ApiNotificationsGet",
        "consumes": [
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "connectorId",
            "in": "query",
            "description": "Connector ID for multi tenanted connections",
            "required": false,
            "type": "string"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer "
          }
        ],
        "responses": {
          "200": {
            "description": "Returned if the request was processed successfully.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ConnectorNotificationModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Acknowledges a notification as having been processed by the connector.",
        "operationId": "_ApiNotificationsPost",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "acknowledge",
            "in": "body",
            "description": "The acknowledgement information.",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConnectorNotificationAcknowledgeModel"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer "
          }
        ],
        "responses": {
          "200": {
            "description": "Returned if the request was processed successfully."
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponseModel"
            }
          },
          "429": {
            "description": "Client Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponseModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "MetaDataModel": {
      "description": "Represents a single metadata entry.",
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "description": "Internal name for the field",
          "type": "string"
        },
        "type": {
          "description": "Type of the metadata entry.",
          "type": "string"
        },
        "value": {
          "description": "Value of the metadata entry.",
          "type": "string"
        }
      }
    },
...
    "ErrorResponseModel": {
      "description": "Top level error response object.\r\nDefined by Microsoft REST API design guidelines:\r\nhttps://github.com/Microsoft/api-guidelines/blob/master/Guidelines.md#710-response-formats",
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/ErrorModel",
          "description": ""
        }
      }
    },
    "ErrorModel": {
      "description": "Error object.\r\nDefined by Microsoft REST API design guidelines:\r\nhttps://github.com/Microsoft/api-guidelines/blob/master/Guidelines.md#710-response-formats",
      "type": "object",
      "properties": {
        "type": {
          "enum": [
            "general",
            "aggregate",
            "validation",
            "partialSuccess"
          ],
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "messageCode": {
          "type": "string"
        },
        "severity": {
          "enum": [
            "critical",
            "error",
            "warning",
            "informational"
          ],
          "type": "string"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "target": {
          "type": "string"
        },
        "innerError": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorModel"
          }
        }
      }
    },
...
    "ConnectorConfigModel": {
      "description": "Connector configuration. \r\nConnector configuration models are designed to be \"one size fits all\", \r\nthat is, this model schema is designed to be able to be used for all \r\nconnector types.",
      "required": [
        "status",
        "displayName",
        "hasSubmittedData"
      ],
      "type": "object",
      "properties": {
        "id": {
          "description": "ID of this Connector instance",
          "type": "string"
        },
        "transactionId": {
          "description": "Current Transaction ID on this object",
          "type": "string"
        },
        "connectorTypeId": {
          "description": "ID of the ConnectorType for this instance",
          "type": "string"
        },
        "status": {
          "description": "\"Enabled\" if the connector is enabled, \"Disabled\" if disabled,\r\n\"Error\" if in error state\r\nMust be set to \"Disabled\" when connector is being created\r\nCan be set to \"Enabled\", if connector is authorized in content source\r\nand aggregation level and versioning level are provided",
          "type": "string"
        },
        "statusCode": {
          "description": "String containing general reason for last connector error state",
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "description": "Date and time that this connector instance was created",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "description": "Date and time that this connector instance was last modified",
          "type": "string"
        },
        "createdBy": {
          "description": "Username of the user that created this connector instance",
          "type": "string"
        },
        "modifiedBy": {
          "description": "Username of the user that last modified this connector instance",
          "type": "string"
        },
        "tenantId": {
          "description": "The ID of the Records365 vNext tenant that owns this connector instance",
          "type": "string"
        },
        "tenantDomainName": {
          "description": "The domain name of the Records365 vNext tenant that owns this connector instance",
          "type": "string"
        },
        "originatingOrganization": {
          "description": "Originating organization field to map with the name of the organization the connector is connecting to",
          "type": "string"
        },
        "enabledHistory": {
          "description": "True if this connector has ever been enabled in its lifetime, false otherwise. This is a read-only property",
          "type": "string"
        },
        "displayName": {
          "description": "The display name for this connector instance",
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "hasSubmittedData": {
          "description": "Whether the connector has submitted any items",
          "type": "boolean"
        },
        "properties": {
          "description": "Custom properties for the connector instance",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MetaDataModel"
          }
        },
        "clientId": {
          "description": "Client ID of the OAuth 2.0 Application (e.g., \r\nan Azure AD Application) that represents this Connector.\r\nIs used instead of the Connector Type's Client ID if present",
          "type": "string"
        },
        "protectionEnabled": {
          "description": "Whether or not binary protection is enabled for this connector",
          "type": "boolean"
        },
        "filters": {
          "$ref": "#/definitions/FiltersModel",
          "description": "Filters records based on Excluded and Included properties"
        }
      }
    },
    "FiltersModel": {
      "description": "Filters records based on Excluded and Included properties.",
      "type": "object",
      "properties": {
        "included": {
          "$ref": "#/definitions/SearchTreeNodeModel"
        },
        "excluded": {
          "$ref": "#/definitions/SearchTreeNodeModel"
        },
        "includedExpression": {
          "description": "Holds user's input included expression",
          "type": "string"
        },
        "excludedExpression": {
          "description": "Holds user's input excluded expression",
          "type": "string"
        }
      }
    },
    "SearchTreeNodeModel": {
      "description": "Groups multiple filters to allow for complex searching",
      "required": [
        "boolOperator"
      ],
      "type": "object",
      "properties": {
        "boolOperator": {
          "description": "Boolean operators for building the search expression\r\nMust be \"OR\" or \"AND\"",
          "maxLength": 3,
          "type": "string"
        },
        "children": {
          "description": "Children search tree nodes for the search expression",
          "type": "array",
          "items": {
            "$ref": "#/definitions/SearchTreeNodeModel"
          }
        },
        "searchTerm": {
          "$ref": "#/definitions/SearchTermModel",
          "description": "Search term including the field name, value and operator"
        }
      }
    },
    "SearchTermModel": {
      "description": "Represents a single search per field\r\neg. ItemTypeId:'0' would have FieldName=\"ItemTypeId\", Operator=\"Equal\" and Value=\"0\"",
      "required": [
        "fieldName"
      ],
      "type": "object",
      "properties": {
        "fieldName": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "fieldType": {
          "type": "string"
        },
        "operator": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "fieldValue": {
          "type": "string"
        },
        "categoricalValueType": {
          "type": "string"
        },
        "searchContextIdentifier": {
          "type": "string"
        }
      }
    },
    "ItemSubmissionOutputModel": {
      "description": "",
      "required": [
        "externalId",
        "connectorId",
        "title",
        "author",
        "sourceLastModifiedDate",
        "sourceLastModifiedBy",
        "sourceCreatedBy",
        "sourceCreatedDate",
        "contentVersion",
        "location",
        "mediaType",
        "parentExternalId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "description": "Internal unique ID for the item",
          "type": "string"
        },
        "itemType": {
          "description": "Defines the type of the item",
          "type": "string"
        },
        "itemNumber": {
          "description": "The internal Records365 vNext item number given to the submitted item",
          "type": "string"
        },
        "format": {
          "description": "The mime format of the submitted item as determined from the file extension",
          "type": "string"
        },
        "lastModifiedDate": {
          "format": "date-time",
          "description": "The date and time of when the item was last modified in Records365 vNext",
          "type": "string"
        },
        "lastModifiedBy": {
          "description": "The identifier of who or what last modified the item in Records365 vNext",
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "description": "The date and time of when the item was created in Records365 vNext",
          "type": "string"
        },
        "createdBy": {
          "description": "The identifier of who or what created the record in Records365 vNext",
          "type": "string"
        },
        "contentSource": {
          "description": "The name of content source from which the item was submitted",
          "type": "string"
        },
        "connectorDisplayName": {
          "description": "The name of the connector that submitted the item",
          "type": "string"
        },
        "isVitalRecord": {
          "description": "True if the item is a vital record, false otherwise",
          "type": "boolean"
        },
        "originatingOrganization": {
          "description": "The name of the organization that submitted the item",
          "type": "string"
        },
        "previousDisposalAction": {
          "description": "Previous disposal action taken on this item",
          "type": "string"
        },
        "previousDisposalDate": {
          "format": "date-time",
          "description": "Date of when the previous disposal action was taken on this item",
          "type": "string"
        },
        "previousDisposalBy": {
          "description": "The user name of the user who completed the last disposal action",
          "type": "string"
        },
        "previousDisposalById": {
          "description": "For auditing purposes, the ID of the user who completed the last disposal action",
          "type": "string"
        },
        "nextDisposalAction": {
          "description": "Next disposition phase action to be taken on this item",
          "type": "string"
        },
        "nextDisposalDate": {
          "format": "date-time",
          "description": "Next disposition phase action date",
          "type": "string"
        },
        "currentDisposalStatus": {
          "description": "The current disposal status of the item, e.g. \"Destroyed\", \"Disposal In-Progress\", \"Error\" etc",
          "type": "string"
        },
        "sourceProperties": {
          "description": "Source properties for the item as populated from the content source",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MetaDataModel"
          }
        },
        "externalId": {
          "description": "The item external ID. This can be anything, but make sure it's unique to the external content source",
          "type": "string"
        },
        "connectorId": {
          "description": "The connector ID. Unique ID of the submitting connector",
          "type": "string"
        },
        "title": {
          "description": "The title of the submitted item",
          "type": "string"
        },
        "author": {
          "description": "The person, office, or designated position responsible for creation or issuance of a document.  \r\nThe author, originator, or creator is usually indicated on the letterhead or by signature.  \r\nFor RMA purposes, the author, originator, or creator may be designated as a person, official title, office symbol, or code",
          "type": "string"
        },
        "mimeType": {
          "description": "The mime type of the submitted item",
          "type": "string"
        },
        "sourceLastModifiedDate": {
          "format": "date-time",
          "description": "The date and time that the item was last modified in the content source. Must be in UTC and ISO-8601 Round-trip format",
          "type": "string"
        },
        "sourceLastModifiedBy": {
          "description": "The name of the user or principal that last modified the item in the content source",
          "type": "string"
        },
        "sourceCreatedBy": {
          "description": "The name of the user or principal that created the item in the content source",
          "type": "string"
        },
        "sourceCreatedDate": {
          "format": "date-time",
          "description": "The date and time that the item was created in the content source. Must be in UTC and ISO-8601 Round-trip format",
          "type": "string"
        },
        "contentVersion": {
          "description": "The version as stored in the content source",
          "type": "string"
        },
        "location": {
          "description": "A pointer to a record???s location.  \r\nExamples: an operating system path and filename, the location cited within a file plan, or the location of a magnetic tape rack",
          "type": "string"
        },
        "mediaType": {
          "description": "The media type, if set to empty, then the media type will default to \"Electronic\"",
          "type": "string"
        },
        "parentExternalId": {
          "description": "A pointer to a record???s aggregation",
          "type": "string"
        },
        "barcodeType": {
          "description": "The barcode type, this can be empty",
          "type": "string"
        },
        "barcodeValue": {
          "description": "The barcode value of the above barcode type",
          "type": "string"
        },
        "correlationId": {
          "description": "An optional ID used to track the item version as it moves through Records365 vNext's processing pipeline",
          "type": "string"
        }
      }
    },
...
    "ConnectorNotificationModel": {
      "description": "A notification sent to the Content Source",
      "type": "object",
      "properties": {
        "id": {
          "description": "Gets or sets unique ID of this notification. The receiver may use\r\nthis value to deduplicate incoming notifications",
          "type": "string"
        },
        "notificationType": {
          "description": "Gets or sets type of notification. Possible values include:\r\n'ItemDestroyed', 'Ping', 'ConnectorConfigAdded',\r\n'ConnectorConfigUpdated', 'ConnectorConfigDeleted',\r\n'GenerateReport'",
          "type": "string"
        },
        "timestamp": {
          "format": "date-time",
          "description": "Gets or sets UTC timestamp that indicates when the event occured in\r\nRecords365 vNext",
          "type": "string"
        },
        "tenantId": {
          "description": "Gets or sets the ID of the Records365 vNext tenant that relates to\r\nthis notification",
          "type": "string"
        },
        "connectorId": {
          "description": "Gets or sets the ID of the connector instance that relates to this\r\nnotification",
          "type": "string"
        },
        "item": {
          "$ref": "#/definitions/ItemSubmissionOutputModel",
          "description": "Gets or sets the item which this notification relates to. Null if\r\nthe notification relates to a connector"
        },
        "connectorConfig": {
          "$ref": "#/definitions/ConnectorConfigModel",
          "description": "Gets or sets the connector configuraiton which this notification\r\nrelates to. Null if the notification relates to an item"
        }
      }
    },
    "ConnectorNotificationAcknowledgeModel": {
      "description": "Information sent by a connector to acknowledge that it has processed a notification.",
      "required": [
        "connectorId",
        "notificationId",
        "processingResult"
      ],
      "type": "object",
      "properties": {
        "connectorId": {
          "description": "Id of the Connector that processed the notification",
          "type": "string"
        },
        "notificationId": {
          "description": "Id of the Notification being acknowledged",
          "type": "string"
        },
        "processingResult": {
          "description": "Processing result of the notification. May be OK, Unknown, ConnectorDisabled, ConnectorNotSubscribed, ConnectorNotReachable, NotificationError",
          "type": "string"
        },
        "connectorStatusMessage": {
          "description": "An optional message from the connector to be added to the items metadata on the \"ConnectorStatusMessage\" field",
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
  }
}